Sorts PokemonDetails screen
Because some of the information present on the details screen is not
pulled from the game master file it is being removed. As a result the
details screen is being reorganized so that we can strike a balance
between functional and clean design.

The Attack component is also being modified because individual DPS
numbers don't matter. Instead we're focusing on energy gain for quick
moves, and giving more breathing room to the damage and cooldown time
sections so they're more readable.