Begin switching to pokemagic pokemon db
In order to update with the latest data from game master we'll be
adopting pokemagic's pokemon database which is already up to date and
provides a lot of the dex functionality that is displayed here. This
will reduce our bundle size since we will no longer bundle two copies of
the pokemon data, and it'll define a single source for that data.

This begins the process which we can adopt incrementally without
breaking the rest of the app's functionality.