Replaces Tools with StrongAgainst
We're also removing the previous StronAgainstList since it was its own
view and we want this one to be embedded into the PokemonInfo screen.
This new list takes into account the movesets (ordered by combo DPS) and
finds the top 20 pokemon (sorted by CP and effectiveness) where that
moveset is most effective.

This also deletes the WeakAgainst/StrongAgainst utils since we are not
using them.